+++ engine-ui/packages/console-ui/src/webapps/rdos/comm/const.js
@@ -166,7 +166,8 @@ export const TASK_TYPE = { // 任务类型
     CARBONSQL: 12, // CarbonSQL
     NOTEBOOK: 13,
     EXPERIMENT: 14,
-    CUBE_KYLIN: 16
+    CUBE_KYLIN: 16,
+    HIVESQL: 17
 }
 
 export const LEARNING_TYPE = {// 深度学习框架
@@ -267,6 +268,7 @@ export const hdfsFieldTypes = [ // hdfs 类型
     'INT',
     'BIGINT',
     'FLOAT',
+    'DECIMAL',
     'DOUBLE',
     'TIMESTAMP',
     'DATE'