+++ engine-ui/packages/console-ui/src/webapps/stream/views/helpDoc/docs.js
@@ -233,5 +233,22 @@ export const sourceFormat = (
 export const analyticalRules = (
     <div>
         <p>自动建表的表名，将按照固定前缀（stream）、源表所属的schema、表名拼接</p>
+        <p>若配置了分表模式，则{'{table}'}将被替换为每个分组名称</p>
+    </div>
+)
+export const multipleTable = (
+    <div>
+        <p>分表模式下，可选择多个表分组，按照不同分组的表，写入不同的目标，实现多对多写入</p>
+    </div>
+)
+
+export const partitionType = (
+    <div>
+        <p>按照天或小时粒度，自动创建分区（字段名pt），并按照数据写入时间，自动写入不同分区</p>
+    </div>
+)
+export const writeTableType = (
+    <div>
+        <p>自动建表的表名，将按照固定前缀（stream）、源表所属的schema、表名拼接，若数据源选择了分表模式，则每个表分组会自动创建一张Hive表</p>
     </div>
 )